+++ AvalonStudio/AvalonStudio.Repository/Logger.cs
@@ -1,5 +1,4 @@
-﻿using AvalonStudio.Utils;
-using NuGet.Common;
+﻿using NuGet.Common;
 
 namespace AvalonStudio.Repositories
 {
@@ -22,33 +21,5 @@ namespace AvalonStudio.Repositories
         public void LogInformationSummary(string data) => $"Infosummary: {data}".Dump();
 
         public void LogErrorSummary(string data) => $"LogErrorSummary: {data }".Dump();
-    }
-
-    public class AvalonConsoleNuGetLogger : ILogger
-    {
-        private IConsole _console;
-
-        public AvalonConsoleNuGetLogger(IConsole console)
-        {
-            _console = console;
-        }
-
-        public void LogDebug(string data) => _console.WriteLine(data);
-
-        public void LogVerbose(string data) => _console.WriteLine(data);
-
-        public void LogInformation(string data) => _console.WriteLine(data);
-
-        public void LogMinimal(string data) => _console.WriteLine(data);
-
-        public void LogWarning(string data) => _console.WriteLine(data);
-
-        public void LogError(string data) => _console.WriteLine(data);
-
-        public void LogSummary(string data) => _console.WriteLine(data);
-
-        public void LogInformationSummary(string data) => _console.WriteLine(data);
-
-        public void LogErrorSummary(string data) => _console.WriteLine(data);
     }
 }
